backend/requirements.txt
fastapi
uvicorn
python-socketio
networkx
pandas
scikit-learn
numpy
faker
joblib
sqlalchemy
